Small changes to Dockerfile
I've added the copy of the `pom.xml` file and `settings.xml` back in.
These are both required for `mvn dependency:go-offline` to run correctly.
I've also removed the policy loading line as the Conjur CE server is not
currently running when this Dockerfile is run.  Setup like this should
happen as part of the test run.