Remove maven-exec, unused dependency
maven-exec is incorrectly imported as a dependency for
the jsonbodyreader. However, this was due to that class
accidentally importing a third-party `Collections`, rather than
the Java Collections.

Using  mvn dependency:analyze-only, other unused dependencies
were removed as well. junit was added, as it was a used and undeclared
dependency.

Further, `Endpoints.java` did not fully implement `Serializable`, which
was a minor error which has been resolved.

Additionally, our test directories needed to be fixed to
use the correct classpath for the project.